# doc/api.yaml
...
openapi: 3.0.0 
info:  
  title: WasaPhoto 
  description:  |  
    This API allows you to manage the 
    data available in the project 
  version: 0.0.1 
servers:
  - url: http://localhost:3000
paths:  
...
  /users/{userId}/banUser/{userId2}:

     parameters:  
      - $ref: "#/components/parameters/userId" 
      - $ref: "#/components/parameters/userId2" 

     put: 
          tags: ["banUser"] 
          summary: Ban another user 
          description: | 
            If the user identified by userId or userId2 does not exist, error is returned. 
            If the user identified by userId or userId2 exists and user isn't banned by user2, user identified by userId ban user identified by userId2. 
          operationId: banUser 
          security: 
            - BearerAuth: [] 

          responses:  
            '200': 
              description: User2 banned
              content: 
                application/json: 
                  schema: 
                    $ref: '#/components/schemas/scUserName' 
                  example:
                    username: "Luigi11" 

            '400': { $ref: "#/components/responses/UserIdNotFound" }            
            '403': {$ref: "#/components/responses/UserIdBanned"}
            '404': {$ref: "#/components/responses/UserId2NotFound"}                 
            '405': {$ref: "#/components/responses/userId2Banned"} 
            '500': {$ref: "#/components/responses/ErrorServerExecution"}   
            '502': {$ref: "#/components/responses/FailedAutentication"}  

     delete:
          tags: ["unbanUser"] 
          summary: Unban another user 
          description: | 
            If the user identified by userId or userId2 does not exist or user doesn't ban userId2, error is returned. 
            If user banned userId2, user identified by userId unabanned user identified by userId2. 
          operationId: unbanUser 
          security: 
            - BearerAuth: [] 

          responses:  
            '200': 
              description: User2 unbanned
              content: 
                application/json: 
                  schema: 
                    $ref: '#/components/schemas/scUserName' 
                  example:
                    username: "Luigi11"  


            '400': { $ref: "#/components/responses/UserIdNotFound" }            
            '403': {$ref: "#/components/responses/UserIdBanned"}
            '404': {$ref: "#/components/responses/UserId2NotFound"}                 
            '406': {$ref: "#/components/responses/userNotBanned"}
            '500': {$ref: "#/components/responses/ErrorServerExecution"}   
            '502': {$ref: "#/components/responses/FailedAutentication"}  

            
     get:
      tags: ["checkBan"] 
      summary: Check if user ban another user 
      description: | 
            If the user identified by userId or userId2 does not exist error is returned. 
            If user banned userId2, the id which identifies the userId2 is returned  
      operationId: checkBan 
      security: 
          - BearerAuth: [] 

      responses:  
            '200': 
              description: UserId2 wasa banned
              content: 
                application/json: 
                  schema: 
                      $ref: "#/components/schemas/userId"
                  example:
                    userId: 111112


            '400': { $ref: "#/components/responses/UserIdNotFound" }            
            '403': {$ref: "#/components/responses/UserIdBanned"}
            '404': {$ref: "#/components/responses/UserId2NotFound"}                 
            '406': {$ref: "#/components/responses/userNotBanned"}
            '500': {$ref: "#/components/responses/ErrorServerExecution"} 
            '502': {$ref: "#/components/responses/FailedAutentication"}  
  /photo/{photoId}/like/{userId}:
    parameters:
      - $ref:  '#/components/parameters/photoId'
      - $ref: '#/components/parameters/userId'

    put:
          tags: ["likePhoto"] 
          summary: Like a post
          description: | 
            If the user identified by userId or photo identified by phId don't exist or one user is banned, error is returned. 
            If the user identified by userId and photo identified by phId exists and user can see the photo , the user likes the photo indicated by phId. 
          operationId: likePhoto 
          security: 
            - BearerAuth: []

          responses:
            '200': 
              description: Liked
              content: 
                application/json: 
                  schema: 
                    $ref: '#/components/schemas/errorSchema' 
                  example:
                    result: "Done"     

            '400': { $ref: "#/components/responses/UserIdNotFound" }    
            '407': {$ref: "#/components/responses/photoNotFound"} 
            '403': {$ref: "#/components/responses/UserIdBanned"}
            '405': {$ref: "#/components/responses/userId2Banned"} 
            '500': {$ref: "#/components/responses/ErrorServerExecution"}  
            '502': {$ref: "#/components/responses/FailedAutentication"}  


    delete:
          tags: ["unlikePhoto"] 
          summary:  Remove like in a photo 
          description: | 
            If the user identified by userId or photoId  does not exist or user doesn't liked the photo, error is returned. 
            If user had liked the photo and he isn't banned, he removes his likes on the post
          operationId: unlikePhoto 
          security: 
            - BearerAuth: [] 

          responses:  
            '200': 
              description: Remove like in a photo
              content: 
                application/json: 
                  schema: 
                    $ref: '#/components/schemas/errorSchema' 
                  example:
                    result: "Done" 

            '400': { $ref: "#/components/responses/UserIdNotFound" }    
            '403': {$ref: "#/components/responses/UserIdBanned"}
            '405': {$ref: "#/components/responses/userId2Banned"}
            '407': {$ref: "#/components/responses/photoNotFound"} 
            '408': {$ref: "#/components/responses/LikeNotFound"} 
            '500': {$ref: "#/components/responses/ErrorServerExecution"}  
            '502': {$ref: "#/components/responses/FailedAutentication"}  
...
  /users/{userId}/photo:
    parameters:
      - $ref: '#/components/parameters/userId'
    post:
          tags: ["uploadPhoto"] 
          summary:  Upload new photo 
          description: | 
            If the user identified by userId  doesn't exist, error is returned. 
            If the user identified by userId exist, the Photo is uploaded.
          operationId: uploadPhoto 
          security: 
            - BearerAuth: [] 
          requestBody:
           description: The binary image file to upload
           content:
            image/*:
              schema:
                  $ref: "#/components/schemas/image"
          responses:
            '201': 
              description: Photo uploaded
              content: 
                application/json: 
                  schema: 
                    $ref: '#/components/schemas/errorSchema' 
                  example:
                    result: "Done" 

            '400': { $ref: "#/components/responses/UserIdNotFound" }                       
            '410': { $ref: "#/components/responses/InvalidFormat" }   
            '500': {$ref: "#/components/responses/ErrorServerExecution"}  
            '502': {$ref: "#/components/responses/FailedAutentication"}  

          
    get:
      tags: ["getUserPhotos"]
      summary: Returns user photos
      description: |
        If the user identified by userId  doesn't exist, error is returned. 
        If the user identified by userId exist, returns the list of photos uploaded by a user
      operationId: getUserPhotos
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Returns the user photos list
          content:
            application/json:
              schema:
                type: array 
                minItems: 0
                maxItems: 100
                description:  Various posts of the user 
                items:
                  $ref: '#/components/schemas/userPhotoStream'
              example:
                - photoId: 12
                  userId: 3
                  name: "Mario_1"
                  likes: 109
                  comments: 124
                  upladTime: "2023-02-12T12:00:00Z"
                - photoId: 52
                  userId: 3
                  name: "Mario_1"
                  likes: 10
                  comments: 18
                  upladTime: "2023-03-18T12:00:00Z"

        '400': { $ref: "#/components/responses/UserIdNotFound" }    
        '403': {$ref: "#/components/responses/UserIdBanned"}
        '404': {$ref: "#/components/responses/UserId2NotFound"}
        '405': {$ref: "#/components/responses/userId2Banned"}     
        '500': {$ref: "#/components/responses/ErrorServerExecution"}    
  /user/{userId}/photo/{photoId}:  
      parameters:
        - $ref:  '#/components/parameters/photoId'
        - $ref: '#/components/parameters/userId'
      delete:
          tags: ["deletePhoto"] 
          summary:  Remove photo from his profile 
          description: | 
            If the user identified by userId or photoId  does not exist or user doesn't posted the photo, error is returned. 
            If user had posted the photo and the photoId exists, he removes his  post
          operationId: deletePhoto 
          security: 
            - BearerAuth: [] 

          responses:  
            '200': 
              description: Remove photo
              content: 
                application/json: 
                  schema: 
                    $ref: '#/components/schemas/errorSchema' 
                  example:
                    result: "Done" 

            '401': { $ref: "#/components/responses/UserNotAuthorized" }
            '407': {$ref: "#/components/responses/photoNotFound"}              
            '500': {$ref: "#/components/responses/ErrorServerExecution"}         


      get:
          tags: ["getPhoto"]
          summary:  Return single photo
          description: Return the single requested photo in binary format.
          operationId: getPhoto
          security:
            - BearerAuth: []
          responses:
            '200':
              description: The requested photo in binary format.
              content:
                application/json:
                  schema:
                    $ref: "#/components/schemas/imageD"
                  example:
                    image: "34qfni23"

            '400': { $ref: "#/components/responses/UserIdNotFound" }   
            '403': {$ref: "#/components/responses/UserIdBanned"}
            '404': {$ref: "#/components/responses/UserId2NotFound"} 
            '405': {$ref: "#/components/responses/userId2Banned"} 
            '407': {$ref: "#/components/responses/photoNotFound"}   
            '500': {$ref: "#/components/responses/ErrorServerExecution"}  
...
  /users:
    get:
      tags: ["getUserId"]
      summary: Get userId from his username
      description:  |
        If user identified by userId in authentication doesn't exist or user identified by the username passed doesn't exist, error is returned.
        If user identified by userId in authentication exist and user identified by the username passed exist, the userId of the user is ruterned.
      operationId: getUserId
      security:
        - BearerAuth: []
      parameters:
        - name: username
          in: query
          schema:
            $ref: "#/components/schemas/username"
          required: true
          description: Username of the user to search 
      responses:
          '201': 
            description: User log-in action successful 
            content: 
              application/json: 
                schema: 
                    $ref: "#/components/schemas/userId"
                example:
                  userId: 111112
          '404': { $ref: "#/components/responses/UsernameNotFound" }    
          '400': { $ref: "#/components/responses/UserIdNotFound" }    
          '500': {$ref: "#/components/responses/ErrorServerExecution"} 
          '411': { $ref: "#/components/responses/UsernameNotValid" }
  /photo/{photoId}/likes:
    get:
        tags: ["getPhotoLike"]
        summary: Gets photo likes
        description:  |
          If the photo identified by photoId  does not exist , error is returned. 
          if the user is not banned, it returns the list of photo likes
        operationId: getPhotoLike
        security:
          - BearerAuth: []
        parameters:
          - $ref: '#/components/parameters/photoId'
        responses:
          '200':
            description: Returns the user list
            content:
              application/json:
                schema:
                  type: array
                  minItems: 0
                  maxItems: 100
                  description: An array of users liking the photo.
                  items:
                    $ref: "#/components/schemas/userId"
                example:
                  - userId: 111112
                  - userId: 1452
          '400': { $ref: "#/components/responses/UserIdNotFound" }    
          '403': {$ref: "#/components/responses/UserIdBanned"}
          '405': {$ref: "#/components/responses/userId2Banned"} 
          '407': {$ref: "#/components/responses/photoNotFound"} 
          '500': {$ref: "#/components/responses/ErrorServerExecution"}    
components:  
  responses:
...
    UsernameNotValid:
     description: Username not valid
    FailedAutentication:
      description: Failed authentication
    UsernameNotFound:
      description: Searched username not found
    UserIdNotFound:
      description: User id not found
    UserNotAuthorized:
      description: User not authorized 
    UserIdBanned:
      description: User banned by User2
    userId2Banned:
      description: User banned User2
    UserId2NotFound:
      description: Userid2 not found    
    ErrorServerExecution: 
      description: Error server execution     
...
    userNotBanned:
      description: User wasn't banned 
    photoNotFound:
      description: Photo Id not found
    LikeNotFound:
      description: User didn't like the photo
...
    InvalidFormat:
      description: Invalid file format
...
  schemas: 
    username: 
      type: string 
      description: User's UserName 
      example: "Franchino_12"
      pattern: '^[a-zA-Z0-9_.-]*$' 
      minLength: 6 
      maxLength: 16  
    userId:
      type: object 
      description: User log-in action successful 
      properties: 
        identifier: 
          $ref: '#/components/schemas/identifier'
    identifier: 
      type: integer
      description: User Id 
      example: 1 
...
    photoId:
      type: integer
      description: Photo identifier
      example: 5
    likeNumber: 
      type: integer
      description: Number of likes of the post 
      example: 987
    commentNumber:
      type: integer
      description: Number of comment of the post  
      example: 10000
...
    uploadTime:
      type: string
      format: date-time
      description: Upload time
      minLength: 20
      maxLength: 25
      example: "2022-01-13T12:00:00Z"
    image:
      description: The photo file
      type: string
      format: binary
      minLength: 1
      maxLength: 10485760 
      pattern: "((.|\n)*)" 
    imageD:
      type: object
      description: "Return codified photo"
      properties:
        image:
          $ref: "#/components/schemas/image"
    userPhotoStream:
      type: object
      properties:
        photoId:
          $ref: '#/components/schemas/photoId'
        likes:
          $ref: '#/components/schemas/likeNumber'
        comments: 
          $ref: '#/components/schemas/commentNumber'
        uploadTime:
          $ref: '#/components/schemas/uploadTime'
        userId: 
          $ref: '#/components/schemas/identifier'
        name: 
          $ref: '#/components/schemas/username'
...
    errorSchema: 
      description: Error example
      type: object 
      properties: 
        Message: 
          $ref: "#/components/schemas/errorDescription" 
    errorDescription: 
      type: string  
      description: Status
      example: Success 
      pattern: '^[a-zA-Z0-9]+$' 
      minLength: 4
      maxLength: 50 
...
    scUserName:
      type: object
      description:  "Describe userName" 
      properties:
        username:
          $ref: "#/components/schemas/username"  
...
  parameters:  
    userId: 
      in: path 
      name: userId 
      required: true 
      description: User Id 
      schema: 
        type: integer 
        description: "User Id" 
        example: 86 
    userId2: 
      in: path 
      name: userId2 
      required: true 
      description: Second userId
      schema: 
        type: integer 
        description: "Second userId" 
        example: 100 
    photoId: 
      in: path 
      name: photoId 
      required: true 
      description: Single id photo
      schema: 
        type: integer 
        description: "Single id photo"
        example: 15368 
...
  securitySchemes: 
    BearerAuth: 
      description: login only as your user id 
      type: http 
      scheme: bearer 